Expanded/Corrected prerequisites for installation
during my installation I discovered that while I had git installed (via github client) I couldn't use the pip install line because the PATH variable was not set.  I figured this warrented some kind of mention in the manual.  I'm not sure how PATH variables work on linux or if this is strictly a windows thing.